[FIX] web_editor: reintroduce request_editable event
The commit fixes the display of custom colors in the color picker.

It improves previous fix from #72294, which was not returning all
editables from the editor, thus only the custom colors of the target
were displayed.
For that the request_editable event is reintroduced at the SnippetsMenu
level.

Part of https://github.com/odoo/odoo/pull/72978
task-2476601

X-original-commit: 3c0e9359c410f71051d74cccc592988ea68654a3