[FIX] purchase_stock: add sequence to new view
The commit caf45f0152ce7d50fc3bc79d850f2afe4eba2841 introduced a new
view on model supplier info. The issue is that this new view is meant to
be used on supplier info as One2Many on stock.orderpoint and so contains
some `parent.` relation.

As no sequence has been set. This view replaced the original one in the
Purchase Configuration menu. Leading to some error as `parent.` in not
defined for standalone tree view.

This commit set a sequence on this new view to ensure it is used only on
the orderpoint form view.

closes odoo/odoo#73212

Signed-off-by: William Henrotin <Whenrow@users.noreply.github.com>