[FIX] website: set id in xml for carousel rendered by configurator
Carousels (carousel and quotes_carousel snippets) in pages created by the
configurator are not added through a drag'n'drop. onBuilt and _assignUniqueId
are thefore not called. We set the unique id directly in the xml to make
the carousel slides also work when these snippets are added by the configurator.

task-2518565

closes odoo/odoo#73256

X-original-commit: 1a6626e6789f31ba6cb0d064f9a8685fd24ab7f1
Signed-off-by: Jérémy Kersten (jke) <jke@openerp.com>