[FIX] website_form_project: prevent errors on task creation from a form
In 14.0, when adding a form to a website and selecting "Create a Task" without
an active project, an error pops up because no projects are found.

In this commit, we prevent an error from popping up when adding the form
without projects, by allowing the field to be empty (not required).

task-2580436

closes odoo/odoo#75221

X-original-commit: a5d8648531f9a0e9616f23b66c67ff32fdb0f0ea
Signed-off-by: Romain Derie <rdeodoo@users.noreply.github.com>