[FIX] account: fix partner mapping with notes on reconcile models
On reconcile models, a rule can be set to map a partner to a bstline if
the bstline "Notes" matches a regex. This "Notes" field is actually
the "narration" field inherited from move. In commit 9a1c6d839fce8218212d01de493d784704f8e188
this field is converted from Text to Html. This breaks the regex match.
The fix converts back the field value to text for the regex check.

closes odoo/odoo#73618

Task: 2427089
X-original-commit: 660f9eeb0dd33f9b00cdae1c99d8549facf3dfdd
Signed-off-by: Laurent Smet <smetl@users.noreply.github.com>
Signed-off-by: jbw-odoo <jbw-odoo@users.noreply.github.com>