[FIX] sale: allow internal users to access UTM campaings
Before this commit:
UTM campaigns are accessible to all internal users. But with sale module
installed, few of the fields are added into the `utm.campaign` model and
it's views. Those fields contain information related to quotations and
invoices. So if the internal user who does not have the basic sales rights
tries to access the campaing, an AccessError is raised.

With this commit:
We only show those fields in the views to the users having enough rights
("sales_team.group_sale_salesman") and thus allow other internal users
to access the UTM campaings without AccessError.

Task-2417993

closes odoo/odoo#77427

X-original-commit: dcee8b4e8f8ac96c2e8436d08f64013d3e0041b5
Related: odoo/enterprise#21278
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>