[FW][IMP] test_mail: test models with type do not mess with attachment types
Purpose of this commit is to highlight an issue that may happens easily with
`crm` that is made generic here within `test_mail`.

`crm` alters the context when creating a new record adding in this case
`default_type` to it][1]. The returned record contains that altered context.
his results in other records created from it trying to assign that same default
value for `type`. This is a very common name for fields, and happens to exist
in `ir.attachment` too.

If you create an alias for incoming leads in your DB with default values
`{"type": "lead"}` (something very common) and then an email comes to that
alias that contains an inlined base64 image, the attachment creation process
would simply fail.

Obtained error is ``ValueError: Wrong value for ir.attachment.type: 'lead'`` .

[1]: https://github.com/odoo/odoo/blob/b3581271a186de4b7e1e76fd8881d7bed6845bf6/addons/crm/models/crm_lead.py#L310-L311

X-original-commit: odoo/odoo@98285a2764c9485589833052f2500920aa999537
Part-of: odoo/odoo#76506
Co-authored-by: Thibault Delavallee <tde@odoo.com>