[FIX] mail: dropzone deactivate for text and user's avatar
Before this commit:

Currently, in Discuss, the chat activates the drop zone when try to drag and
drop any user's avatar (or any text). The same thing happens if we do it in
the chat window or chatter.

After this commit:

The chat will not be activated drop zone if the drag is not actually
a file/a document.

Task-2590510

closes odoo/odoo#80458

X-original-commit: bb7f819f153682a47d1d8b7c9f08a159a666b309
Signed-off-by: Sébastien Theys (seb) <seb@odoo.com>