[FIX] im_livechat: add session.js to manifest
Since 3c4bdec26d54d7fe4cef752c20e074cda44c9745, a new session module is
used to access the session info. The legacy session uses also this
module to access the session info. This means that if a module uses the
legacy session, the new session.js needs to be added to the manifest.

closes odoo/odoo#73500

Related: odoo/enterprise#19581
Signed-off-by: Sébastien Theys (seb) <seb@odoo.com>