[FIX] website: restore overlay for timeline cards edition
With [1], not all overlays are displayed anymore but we rather show the
parent one in some cases. The timeline cards were considered as one of
those cases while it should not, as it prevents to use the overlay
buttons allowing to switch the cards order.

[1]: https://github.com/odoo/odoo/commit/b8d6850f3e76dc6c7c15b3156111eac83d7fba58

Discovered while working on task-2431484

closes odoo/odoo#73515

X-original-commit: c6025c2b9191b9e1bc6cbddcca49847c7f423c13
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>