[FIX] web_editor: save zone with proper res_model
Without this commit, the method `saveModifiedImages` did not use
editables zone but rather the whole `$editable`.
Because of that, oeModel and oeId were undefined and saving images
was done without the proper metadata.

That made impossible the retrieval of the images from the method
`ir.qweb.field.image` `from_html` defined by the module `web_unsplash`.

task-2581567

closes odoo/odoo#77248

X-original-commit: 7a1946c87d58950d1a9d204bd85f3ced5bab1a55
Signed-off-by: David Monjoie (dmo) <dmo@odoo.com>